requirements.txt
Django==1.8
argparse==1.2.1
django-angular==0.7.13
django-tastypie==0.12.1
nodeenv==0.13.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
wsgiref==0.1.2
